minor: bug fixes

+++ Drivers/eMMC/Inc/eMMC.hpp
@@ -36,9 +36,9 @@ namespace eMMC {
     struct memoryItemHandler {
         // memoryItemData();
         // explicit memoryItemData(uint32_t newSize);
-        uint32_t size;
-        uint32_t startAddress;
-        uint32_t endAddress;
+        uint64_t size;
+        uint64_t startAddress;
+        uint64_t endAddress;
         memoryItemHandler() : size(0), startAddress(0), endAddress(0) {}
         explicit memoryItemHandler(uint32_t newSize)
             : size(newSize), startAddress(0), endAddress(0) {}
